More bullets

+++ src/deploy-secrets.sh
@@ -176,11 +176,17 @@ fi
 ##
 
 # Delete then update the smdp-cacert
-kubectl delete secret smdp-cacert --namespace="${NAMESPACE}"
+if [[ -n "$(kubectl describe secrets/smdp-cacert --namespace=${NAMESPACE} | grep NotFound)" ]] ; then 
+    kubectl delete secret smdp-cacert --namespace="${NAMESPACE}"
+fi 
+
 kubectl create secret generic smdp-cacert --namespace="${NAMESPACE}" --from-file="${TEMPORARY_ES2PLUS_RETURN_CERT_AND_KEY_FILE}"
 
 # Then do the same for sim manager secrets
-kubectl delete secret simmgprod-secrets --namespace="${NAMESPACE}"
+if [[ -n "$(kubectl describe secrets/simmgr-prod-secrets --namespace=${NAMESPACE} | grep NotFound)" ]] ; then 
+    kubectl delete secret simmg-prod-secrets --namespace="${NAMESPACE}"
+fi 
+
 kubectl create secret generic simmgr-prod-secrets \
          --namespace="${NAMESPACE}" \
          --from-literal dbUser=${DB_USER} \
